[Proxying] Cancel proxied work when the worker thread dies
Update the synchronous proxying APIs to return 0 when attempting to proxy to a
thread that has already exited or been canceled or exits or is canceled before
the proxied work is completed.

To implement cancellation, add an optional `cancel` function pointer to the
tasks queued in `em_task_queue`s, which are the internal thread-specific queues
used in the implementation of `em_proxying_queue`. This `cancel` function
pointer is not controllable directly by users of the API, but is rather used
internally to ensure that synchronous waiters are notified when the thread they
are waiting on dies. This mechanism will scale to other forms of cancellation
notification in the future. For example, it could be used to reject a promise
associated with the proxied work.

Threads were previously notified of new work via postMessage messages that
carried pointers to the task queues to execute. There is no way to synchronously
pump or inspect these pending messages however, and there is no central registry
of all task queues for a thread, so this mechanism afforded no way to discover
pending work that would need to be canceled when a thread exits or is canceled.
To fix this, move the pending messages into userspace by giving each thread a
"mailbox", which is an `em_task_queue` in the pthread struct. Instead of using
`postMessage`, proxying queues now use the thread mailbox API to notify threads
of new work.

Internally, thread mailboxes use the same postMessage-based API that proxying
queues used to use, but this is now an implementation detail of the thread
mailbox rather than of the proxying queues. The key difference is that pending
messages now reside in linear memory, where they can be discovered and canceled
during thread shutdown.

When a thread dies because it exits or is canceled, it "closes" its mailbox by
decrementing a refcount and waiting to observe a refcount of 0. At this point,
the thread mailbox API ensures that no new messages will be enqueued on the
mailbox. The thread then runs the cancellation handlers on each task in its
mailbox. Those tasks each carry a pointer to some task queue owned by a proxying
queue, so their cancellation handlers in turn run the cancellation handlers for
all the tasks in those pointed-to queues.

Task queue cancellation handlers do not handle the case of proxied work that has
been started but not finished when the worker thread dies because that work has
already been removed from the task queues. To cancel in-progress work as well,
update the implementation of synchronous proxying to place `em_proxying_ctx`
objects associated with active work in a thread-local doubly linked list that is
traversed by a registered thread specific data destructor. The destructor takes
care of canceling the active work.

Implements the design sketched out in #18631.